requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
asn1crypto==1.5.1
attrs==25.3.0
blinker==1.9.0
boto3==1.40.28
botocore==1.40.28
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
cryptography==45.0.7
distro==1.9.0
filelock==3.19.1
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.2
narwhals==2.4.0
numpy==2.3.3
openai==1.107.1
packaging==24.2
pandas==2.2.3
pillow==11.3.0
platformdirs==4.4.0
protobuf==5.29.5
pyarrow==21.0.0
pycparser==2.23
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
PyJWT==2.10.1
pyOpenSSL==25.1.0
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.27.1
s3transfer==0.14.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snowflake-connector-python==3.17.3
sortedcontainers==2.4.0
sseclient-py==1.8.0
streamlit==1.43.2
tenacity==9.1.2
toml==0.10.2
tomlkit==0.13.3
tornado==6.5.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
